Address the third round of review comments
- Rename `readBlock` to `getBlock` in the `InputConcept` (and everywhere
  else), and document that it may perform I/O, decompression, etc.
- Require `appendToBlock` to return `void` in the `InputConcept`.
- Move `VectorInput`/`makeVectorInput` from `RunsInputPolicy.h` into
  `test/util/ParallelBlockMergeTestHelpers.h`, as they only serve the tests.
- Explain why `isEquivalent` negates the comparator, and why the loop over
  `remainingChunkSize` in `targetsFromChunkSizes` runs more than once.
- Make a non-null cancellation handle an invariant of the `MergeState`.
- Remove the redundant parentheses in `ad_utility::detail::MoveIf`.
- Explain why `makeRandomRuns` needs two random generators, and implement
  `concatenation` via `views::join`.
